// src/app/generated/dcc/model/quantityDto.ts
/**
 * OpenAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { HybridValues } from './hybridValues';
import { CoverageIntervalDto } from './coverageIntervalDto';
import { DimensionDto } from './dimensionDto';
import { UncertaintyDto } from './uncertaintyDto';
import { RichContentDto } from './richContentDto';
import { XmlValuesDto } from './xmlValuesDto';
import { MethodDto } from './methodDto';


export interface QuantityDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    label?: string;
    name?: LanguageSpecificStringsDto;
    dimension?: DimensionDto;
    quantityTypeName?: string;
    hybridValues?: HybridValues;
    noQuantity?: RichContentDto;
    description?: RichContentDto;
    xmlValues?: XmlValuesDto;
    timestamp?: string;
    usedMethods?: Array<MethodDto>;
    uncertainty?: UncertaintyDto;
    coverageInterval?: CoverageIntervalDto;
}